+++ src/lib/project.ts
@@ -28,6 +28,39 @@ export class DenvigProject {
     return `${globalConfig.codeRootDir}/${this.slug}`
   }
 
+  get packageManagers(): string[] {
+    const rootFiles = this.rootFiles
+    const packageManagers = []
+
+    if (
+      rootFiles.includes('pnpm-lock.yaml') ||
+      rootFiles.includes('package.json')
+    ) {
+      packageManagers.push('pnpm')
+    }
+    if (
+      rootFiles.includes('package-lock.json') ||
+      rootFiles.includes('package.json')
+    ) {
+      packageManagers.push('npm')
+    }
+    if (rootFiles.includes('yarn.lock')) {
+      packageManagers.push('yarn')
+    }
+    if (rootFiles.includes('deno.json') || rootFiles.includes('deno.jsonc')) {
+      packageManagers.push('deno')
+    }
+    if (rootFiles.includes('pyproject.toml')) {
+      packageManagers.push('uv')
+    }
+
+    return packageManagers
+  }
+
+  get primaryPackageManager(): string | null {
+    return this.packageManagers[0] || null
+  }
+
   get dependencies() {
     return detectDependencies(this)
   }